Allow more fine-grained throttling of login attempts
In addition to the 5 attempts every 5 minutes rule, add some long
term rules. Its extraordinarily unlikely that a non-malicious person would
use the wrong password 150 times in a row, so add a rule that you
can't have 150 login fails in a row in 48 hours all from the same
IP address. Also add a rule for rate limitting to a max of 75
login attempts every 20 minutes across all IPs (In case of an
attacker with a large number of IPs). One has to be careful about
the possibility of enabling a DOS attack when doing rate limitting
for all IPs, but when it gets so high as to be 75/20 minutes, the
greater risk is of someone brute forcing.

Bug: T122164

+++ includes/specials/SpecialUserlogin.php
@@ -21,6 +21,7 @@
  * @ingroup SpecialPage
  */
 use MediaWiki\Logger\LoggerFactory;
+use Psr\Log\LogLevel;
 use MediaWiki\Session\SessionManager;
 
 /**
@@ -86,6 +87,11 @@ class LoginForm extends SpecialPage {
 	);
 
 	public $mAbortLoginErrorMsg = null;
+	/**
+	 * @var int How many seconds user is throttled for
+	 * @since 1.27
+	 */
+	public $mThrottleWait = '?';
 
 	protected $mUsername;
 	protected $mPassword;
@@ -745,8 +751,9 @@ class LoginForm extends SpecialPage {
 			return self::NEED_TOKEN;
 		}
 
-		$throttleCount = self::incLoginThrottle( $this->mUsername );
-		if ( $throttleCount === true ) {
+		$throttleCount = self::incrementLoginThrottle( $this->mUsername );
+		if ( $throttleCount ) {
+			$this->mThrottleWait = $throttleCount['wait'];
 			return self::THROTTLED;
 		}
 
@@ -862,9 +869,7 @@ class LoginForm extends SpecialPage {
 			$this->getContext()->setUser( $u );
 
 			// Please reset throttle for successful logins, thanks!
-			if ( $throttleCount ) {
-				self::clearLoginThrottle( $this->mUsername );
-			}
+			self::clearLoginThrottle( $this->mUsername );
 
 			if ( $isAutoCreated ) {
 				// Must be run after $wgUser is set, for correct new user log
@@ -881,31 +886,90 @@ class LoginForm extends SpecialPage {
 	/**
 	 * Increment the login attempt throttle hit count for the (username,current IP)
 	 * tuple unless the throttle was already reached.
+	 *
+	 * @since 1.27 Return value changed.
 	 * @param string $username The user name
-	 * @return bool|int The integer hit count or True if it is already at the limit
+	 * @return bool|array false if below limit or an array if above limit
+	 *   Array contains keys wait, count, and throttleIndex
 	 */
-	public static function incLoginThrottle( $username ) {
+	public static function incrementLoginThrottle( $username ) {
 		global $wgPasswordAttemptThrottle, $wgRequest;
-		$username = trim( $username ); // sanity
+		$username = User::getCanonicalName( $username, 'usable' ) ?: $username;
 
 		$throttleCount = 0;
 		if ( is_array( $wgPasswordAttemptThrottle ) ) {
-			$throttleKey = wfGlobalCacheKey( 'password-throttle', $wgRequest->getIP(), md5( $username ) );
-			$count = $wgPasswordAttemptThrottle['count'];
-			$period = $wgPasswordAttemptThrottle['seconds'];
+			$throttleConfig = $wgPasswordAttemptThrottle;
+			if ( isset( $wgPasswordAttemptThrottle['count'] ) ) {
+				// old style. Convert for backwards compat.
+				$throttleConfig = array( $wgPasswordAttemptThrottle );
+			}
+			foreach ( $throttleConfig as $index => $specificThrottle ) {
+				if ( isset( $specificThrottle['allIPs'] ) ) {
+					$ip = 'All';
+				} else {
+					$ip = $wgRequest->getIP();
+				}
+				$throttleKey = wfGlobalCacheKey( 'password-throttle',
+					$index, $ip, md5( $username )
+				);
+				$count = $specificThrottle['count'];
+				$period = $specificThrottle['seconds'];
 
-			$cache = ObjectCache::getLocalClusterInstance();
-			$throttleCount = $cache->get( $throttleKey );
-			if ( !$throttleCount ) {
-				$cache->add( $throttleKey, 1, $period ); // start counter
-			} elseif ( $throttleCount < $count ) {
-				$cache->incr( $throttleKey );
-			} elseif ( $throttleCount >= $count ) {
-				return true;
+				$cache = ObjectCache::getLocalClusterInstance();
+				$throttleCount = $cache->get( $throttleKey );
+				if ( !$throttleCount ) {
+					$cache->add( $throttleKey, 1, $period ); // start counter
+				} elseif ( $throttleCount < $count ) {
+					$cache->incr( $throttleKey );
+				} elseif ( $throttleCount >= $count ) {
+					$logMsg = 'Login attempt rejected because logins to '
+						. '{acct} from IP {ip} have been throttled for '
+						. '{period} seconds due to {count} failed attempts';
+					// If we are hitting a throttle for >= 50 attempts,
+					// it is much more likely to be an attack than someone
+					// simply forgetting their password, so log it at a
+					// higher level.
+					$level = $count >= 50 ? LogLevel::WARNING : LogLevel::INFO;
+					// It should be noted that once the throttle is hit,
+					// every attempt to login will generate the log message
+					// until the throttle expires, not just the attempt that
+					// puts the throttle over the top.
+					LoggerFactory::getInstance( 'password-throttle' )->log(
+						$level,
+						$logMsg,
+						array(
+							'ip' => $ip,
+							'period' => $period,
+							'acct' => $username,
+							'count' => $count,
+							'throttleIdentifier' => $index,
+							'method' => __METHOD__
+						)
+					);
+
+					return array(
+						'throttleIndex' => $index,
+						'wait' => $period,
+						'count' => $count
+					);
+				}
 			}
 		}
+		return false;
+	}
 
-		return $throttleCount;
+	/**
+	 * Increment the login attempt throttle hit count for the (username,current IP)
+	 * tuple unless the throttle was already reached.
+	 *
+	 * @deprecated Use LoginForm::incrementLoginThrottle instead
+	 * @param string $username The user name
+	 * @return bool|int true if above throttle, or 0 (prior to 1.27, returned current count)
+	 */
+	public static function incLoginThrottle( $username ) {
+		wfDeprecated( __METHOD__, "1.27" );
+		$res = self::incrementLoginThrottle( $username );
+		return is_array( $res ) ? true : 0;
 	}
 
 	/**
@@ -914,11 +978,27 @@ class LoginForm extends SpecialPage {
 	 * @return void
 	 */
 	public static function clearLoginThrottle( $username ) {
-		global $wgRequest;
-		$username = trim( $username ); // sanity
+		global $wgRequest, $wgPasswordAttemptThrottle;
+		$username = User::getCanonicalName( $username, 'usable' ) ?: $username;
 
-		$throttleKey = wfGlobalCacheKey( 'password-throttle', $wgRequest->getIP(), md5( $username ) );
-		ObjectCache::getLocalClusterInstance()->delete( $throttleKey );
+		if ( is_array( $wgPasswordAttemptThrottle ) ) {
+			$throttleConfig = $wgPasswordAttemptThrottle;
+			if ( isset( $wgPasswordAttemptThrottle['count'] ) ) {
+				// old style. Convert for backwards compat.
+				$throttleConfig = array( $wgPasswordAttemptThrottle );
+			}
+			foreach ( $throttleConfig as $index => $specificThrottle ) {
+				if ( isset( $specificThrottle['allIPs'] ) ) {
+					$ip = 'All';
+				} else {
+					$ip = $wgRequest->getIP();
+				}
+				$throttleKey = wfGlobalCacheKey( 'password-throttle', $index,
+					$ip, md5( $username )
+				);
+				ObjectCache::getLocalClusterInstance()->delete( $throttleKey );
+			}
+		}
 	}
 
 	/**
@@ -977,7 +1057,7 @@ class LoginForm extends SpecialPage {
 	}
 
 	function processLogin() {
-		global $wgLang, $wgSecureLogin, $wgPasswordAttemptThrottle, $wgInvalidPasswordReset;
+		global $wgLang, $wgSecureLogin, $wgInvalidPasswordReset;
 
 		$cache = ObjectCache::getLocalClusterInstance();
 		$authRes = $this->authenticateUserData();
@@ -999,10 +1079,9 @@ class LoginForm extends SpecialPage {
 				self::clearLoginToken();
 
 				// Reset the throttle
-				$request = $this->getRequest();
-				$key = wfGlobalCacheKey( 'password-throttle', $request->getIP(), md5( $this->mUsername ) );
-				$cache->delete( $key );
+				self::clearLoginThrottle( $this->mUsername );
 
+				$request = $this->getRequest();
 				if ( $this->hasSessionCookie() || $this->mSkipCookieCheck ) {
 					/* Replace the language object to provide user interface in
 					 * correct language immediately on this first page load.
@@ -1079,8 +1158,7 @@ class LoginForm extends SpecialPage {
 			case self::THROTTLED:
 				$error = $this->mAbortLoginErrorMsg ?: 'login-throttled';
 				$this->mainLoginForm( $this->msg( $error )
-					->params( $this->getLanguage()->formatDuration( $wgPasswordAttemptThrottle['seconds'] ) )
-					->text()
+					->durationParams( $this->mThrottleWait )->text()
 				);
 				break;
 			case self::USER_BLOCKED:
